Revert "fix: Re-organise k8s clients"
This reverts commit 57ca756d5d5b150d7c33c3dab69af831ccf151fe.
