+++ components/renku_data_services/k8s/client_interfaces.py
@@ -1,0 +1,116 @@
+"""Required interfaces for k8s clients."""
+
+from __future__ import annotations
+
+from collections.abc import AsyncIterable
+from typing import Protocol
+
+from renku_data_services.k8s.constants import ClusterId
+from renku_data_services.k8s.models import (
+    DeletePropagationPolicy,
+    K8sObject,
+    K8sObjectFilter,
+    K8sObjectMeta,
+    K8sPatch,
+    K8sPatches,
+    K8sPriorityClass,
+    K8sResourceQuota,
+    K8sSecret,
+)
+
+
+class ResourceQuotaClient(Protocol):
+    """Methods to manipulate ResourceQuota kubernetes resources."""
+
+    def list_resource_quota(
+        self, label_selector: dict[str, str], cluster_id: ClusterId
+    ) -> AsyncIterable[K8sResourceQuota]:
+        """List resource quotas."""
+        ...
+
+    async def read_resource_quota(self, name: str, cluster_id: ClusterId) -> K8sResourceQuota:
+        """Get a resource quota."""
+        ...
+
+    async def create_resource_quota(self, quota: K8sPatch, cluster_id: ClusterId) -> K8sResourceQuota:
+        """Create a resource quota."""
+        ...
+
+    async def delete_resource_quota(self, name: str, cluster_id: ClusterId) -> None:
+        """Delete a resource quota."""
+        ...
+
+    async def patch_resource_quota(self, name: str, patch: K8sPatches, cluster_id: ClusterId) -> K8sResourceQuota:
+        """Update a resource quota."""
+        ...
+
+
+class SecretClient(Protocol):
+    """Methods to manipulate Secret kubernetes resources."""
+
+    async def get_secret(self, secret: K8sObjectMeta) -> K8sSecret | None:
+        """Get a secret."""
+        ...
+
+    async def create_secret(self, secret: K8sSecret) -> K8sSecret:
+        """Create a secret."""
+        ...
+
+    async def patch_secret(self, secret: K8sObjectMeta, patch: K8sPatches) -> K8sSecret:
+        """Patch an existing secret."""
+        ...
+
+    async def delete_secret(self, secret: K8sObjectMeta) -> None:
+        """Delete a secret."""
+        ...
+
+
+class PriorityClassClient(Protocol):
+    """Methods to manipulate kubernetes Priority Class resources."""
+
+    async def create_priority_class(self, priority_class: K8sPriorityClass) -> K8sPriorityClass:
+        """Create a priority class."""
+        ...
+
+    async def read_priority_class(self, meta: K8sObjectMeta) -> K8sPriorityClass | None:
+        """Retrieve a priority class."""
+        ...
+
+    async def delete_priority_class(
+        self,
+        meta: K8sObjectMeta,
+        propagation_policy: DeletePropagationPolicy = DeletePropagationPolicy.foreground,
+    ) -> None:
+        """Delete a priority class."""
+        ...
+
+
+class K8sClient(Protocol):
+    """Methods to manipulate resources on a Kubernetes cluster."""
+
+    async def create(self, obj: K8sObject, refresh: bool) -> K8sObject:
+        """Create the k8s object."""
+        ...
+
+    async def patch(self, meta: K8sObjectMeta, patch: K8sPatches) -> K8sObject:
+        """Patch a k8s object.
+
+        If the patch is a list we assume that we have a rfc6902 json patch like
+        `[{ "op": "add", "path": "/a/b/c", "value": [ "foo", "bar" ] }]`.
+        If the patch is a dictionary then it is considered to be a rfc7386 json merge patch.
+        """
+        ...
+
+    async def delete(
+        self, meta: K8sObjectMeta, propagation_policy: DeletePropagationPolicy = DeletePropagationPolicy.foreground
+    ) -> None:
+        """Delete a k8s object."""
+        ...
+
+    async def get(self, meta: K8sObjectMeta) -> K8sObject | None:
+        """Get a specific k8s object, None is returned if the object does not exist."""
+        ...
+
+    def list(self, _filter: K8sObjectFilter) -> AsyncIterable[K8sObject]:
+        """List all k8s objects."""
+        ...
