--- components/renku_data_services/k8s/clients/__init__.py
@@ -1,1 +1,0 @@
-"""Kubernetes library client wrappers."""
