+++ test/components/renku_data_services/k8s/test_models.py
@@ -3,9 +3,8 @@
 from box import Box
 from ulid import ULID
 
-from renku_data_services.k8s.clients.secret import K8sSecret
 from renku_data_services.k8s.constants import ClusterId
-from renku_data_services.k8s.models import GVK, K8sObject
+from renku_data_services.k8s.models import GVK, K8sObject, K8sSecret
 
 
 def test_k8s_object_not_render_manifest():
