+++ test/utils.py
@@ -28,9 +28,11 @@ from renku_data_services.data_api.dependencies import DependencyManager
 from renku_data_services.data_connectors.db import DataConnectorRepository, DataConnectorSecretRepository
 from renku_data_services.db_config.config import DBConfig
 from renku_data_services.git.gitlab import DummyGitlabAPI
-from renku_data_services.k8s.clients.core import K8sClusterClientsPool
-from renku_data_services.k8s.clients.priority_class import K8sPriorityClassClient
-from renku_data_services.k8s.clients.resource_quota import K8sResourceQuotaClient
+from renku_data_services.k8s.clients import (
+    K8sClusterClientsPool,
+    K8sPriorityClassClient,
+    K8sResourceQuotaClient,
+)
 from renku_data_services.k8s.config import KubeConfigEnv, get_clusters
 from renku_data_services.k8s.db import K8sDbCache
 from renku_data_services.message_queue.db import ReprovisioningRepository
